changed update and delete methods in repositories? now they throws exceptions

+++ src/main/java/repository/computer/impl/ComputerRepositoryImpl.java
@@ -1,11 +1,14 @@
 package repository.computer.impl;
 
 
+import Exeptions.computer.ComputerDeleteException;
+import Exeptions.computer.ComputerUpdateException;
 import model.computer.Computer;
 import org.hibernate.Session;
 import repository.computer.ComputerRepository;
 import repository.sessionFactory.SessionFactoryAccess;
 
+import javax.persistence.OptimisticLockException;
 import java.util.List;
 
 public class ComputerRepositoryImpl implements ComputerRepository {
@@ -20,17 +23,24 @@ public class ComputerRepositoryImpl implements ComputerRepository {
     }
 
     @Override
-    public Computer update(Computer computer) {
+    public Computer update(Computer computer) throws ComputerUpdateException {
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.update(computer);
-        session.getTransaction().commit();
+        try {
+            session.getTransaction().commit();
+        }catch (OptimisticLockException e){
+            throw new ComputerUpdateException("Error while updating Computer object");
+        }
         session.close();
         return computer;
     }
 
     @Override
-    public Computer delete(Computer computer) {
+    public Computer delete(Computer computer) throws ComputerDeleteException{
+        if (findById(computer.getId()) == null){
+            throw new ComputerDeleteException("There is no computer object with such an id");
+        }
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.delete(computer);
