+++ src/main/java/repository/display/impl/DisplayRepositoryImpl.java
@@ -1,10 +1,13 @@
 package repository.display.impl;
 
+import Exeptions.display.DisplayDeleteException;
+import Exeptions.display.DisplayUpdateException;
 import model.display.Display;
 import org.hibernate.Session;
 import repository.display.DisplayRepository;
 import repository.sessionFactory.SessionFactoryAccess;
 
+import javax.persistence.OptimisticLockException;
 import java.util.List;
 
 public class DisplayRepositoryImpl implements DisplayRepository{
@@ -19,17 +22,24 @@ public class DisplayRepositoryImpl implements DisplayRepository{
     }
 
     @Override
-    public Display update(Display display) {
+    public Display update(Display display) throws DisplayUpdateException {
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.update(display);
-        session.getTransaction().commit();
+        try {
+            session.getTransaction().commit();
+        } catch (OptimisticLockException e){
+            throw new DisplayUpdateException("Error while updating Display object");
+        }
         session.close();
         return display;
     }
 
     @Override
-    public Display delete(Display display) {
+    public Display delete(Display display) throws DisplayDeleteException {
+        if (findById(display.getId()) == null){
+            throw new DisplayDeleteException("There is no Display object with such an id");
+        }
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.delete(display);
