+++ src/main/java/repository/operationSystem/impl/OperationSystemRepositoryImpl.java
@@ -1,10 +1,15 @@
 package repository.operationSystem.impl;
 
+import Exeptions.noteBook.NoteBookDeleteException;
+import Exeptions.noteBook.NoteBookUpdateException;
+import Exeptions.operationSystem.OperationSystemDeleteException;
+import Exeptions.operationSystem.OperationSystemUpdateException;
 import model.operationSystem.OperationSystem;
 import org.hibernate.Session;
 import repository.operationSystem.OperationSystemRepository;
 import repository.sessionFactory.SessionFactoryAccess;
 
+import javax.persistence.OptimisticLockException;
 import java.util.List;
 
 public class OperationSystemRepositoryImpl implements OperationSystemRepository{
@@ -19,17 +24,24 @@ public class OperationSystemRepositoryImpl implements OperationSystemRepository{
     }
 
     @Override
-    public OperationSystem update(OperationSystem operationSystem) {
+    public OperationSystem update(OperationSystem operationSystem) throws OperationSystemUpdateException {
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.update(operationSystem);
-        session.getTransaction().commit();
+        try {
+            session.getTransaction().commit();
+        } catch (OptimisticLockException e) {
+            throw new OperationSystemUpdateException("Error while updating OperationSystem object");
+        }
         session.close();
         return operationSystem;
     }
 
     @Override
-    public OperationSystem delete(OperationSystem operationSystem) {
+    public OperationSystem delete(OperationSystem operationSystem) throws OperationSystemDeleteException {
+        if (findById(operationSystem.getId()) == null) {
+            throw new OperationSystemDeleteException("There is no OperationSystem object with such an id");
+        }
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.delete(operationSystem);
