+++ src/main/java/repository/randomAccessMemory/impl/RandomAccessMemoryRepositoryImpl.java
@@ -1,10 +1,13 @@
 package repository.randomAccessMemory.impl;
 
+import Exeptions.randomAccessMemory.RandomAccessMemoryDeleteException;
+import Exeptions.randomAccessMemory.RandomAccessMemoryUpdateException;
 import model.randomAccesMemory.RandomAccessMemory;
 import org.hibernate.Session;
 import repository.randomAccessMemory.RandomAccessMemoryRepository;
 import repository.sessionFactory.SessionFactoryAccess;
 
+import javax.persistence.OptimisticLockException;
 import java.util.List;
 
 public class RandomAccessMemoryRepositoryImpl implements RandomAccessMemoryRepository {
@@ -19,17 +22,24 @@ public class RandomAccessMemoryRepositoryImpl implements RandomAccessMemoryRepos
     }
 
     @Override
-    public RandomAccessMemory update(RandomAccessMemory randomAccessMemory) {
+    public RandomAccessMemory update(RandomAccessMemory randomAccessMemory) throws RandomAccessMemoryUpdateException {
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.update(randomAccessMemory);
-        session.getTransaction().commit();
+        try {
+            session.getTransaction().commit();
+        } catch (OptimisticLockException e) {
+            throw new RandomAccessMemoryUpdateException("Error while updating RandomAccessMemory object");
+        }
         session.close();
         return randomAccessMemory;
     }
 
     @Override
-    public RandomAccessMemory delete(RandomAccessMemory randomAccessMemory) {
+    public RandomAccessMemory delete(RandomAccessMemory randomAccessMemory) throws RandomAccessMemoryDeleteException {
+        if (findById(randomAccessMemory.getId()) == null) {
+            throw new RandomAccessMemoryDeleteException("There is no RandomAccessMemory object with such an id");
+        }
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.delete(randomAccessMemory);
