+++ src/main/java/repository/user/impl/UserRepositoryImpl.java
@@ -1,11 +1,14 @@
 package repository.user.impl;
 
+import Exeptions.User.UserDeleteException;
+import Exeptions.User.UserUpdateException;
 import model.computer.Computer;
 import model.user.User;
 import org.hibernate.Session;
 import repository.sessionFactory.SessionFactoryAccess;
 import repository.user.UserRepository;
 
+import javax.persistence.OptimisticLockException;
 import java.util.List;
 
 public class UserRepositoryImpl implements UserRepository{
@@ -20,17 +23,24 @@ public class UserRepositoryImpl implements UserRepository{
     }
 
     @Override
-    public User update(User user) {
+    public User update(User user) throws UserUpdateException {
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.update(user);
-        session.getTransaction().commit();
+        try {
+            session.getTransaction().commit();
+        } catch (OptimisticLockException e){
+            throw new UserUpdateException("Error while updating User object");
+        }
         session.close();
         return user;
     }
 
     @Override
-    public User delete(User user) {
+    public User delete(User user) throws UserDeleteException {
+        if (findById(user.getLogIn()) == null){
+            throw new UserDeleteException("There is no User object with such an id");
+        }
         Session session = SessionFactoryAccess.getSessionFactory().openSession();
         session.beginTransaction();
         session.delete(user);
